+++ src/main/java/fuzs/betteranimationscollection/client/renderer/entity/model/ZombieKneesModel.java
@@ -13,10 +13,10 @@ public class ZombieKneesModel<T extends ZombieEntity> extends ZombieModel<T> {
 
         super(modelSizeIn, shortTexture);
 
-        this.rightLeg = BipedKneesModel.getHalfLeg(this, modelSizeIn, -1.9F, 0.0F, 0, 16, false);
-        this.leftLeg = BipedKneesModel.getHalfLeg(this, modelSizeIn, 1.9F, 0.0F, 0, 16, true);
-        this.rightLowerLeg = BipedKneesModel.getHalfLeg(this, modelSizeIn, 0.0F, -6.0F, 0, 22, false);
-        this.leftLowerLeg = BipedKneesModel.getHalfLeg(this, modelSizeIn, 0.0F, -6.0F, 0, 22, true);
+        this.rightLeg = BipedKneesModel.makeHalfLeg(this, modelSizeIn, -1.9F, 0.0F, 0.0F, 0, 16, false);
+        this.leftLeg = BipedKneesModel.makeHalfLeg(this, modelSizeIn, 1.9F, 0.0F, 0.0F, 0, 16, true);
+        this.rightLowerLeg = BipedKneesModel.makeHalfLeg(this, modelSizeIn, 0.0F, -6.0F, -2.0F, 0, 22, false);
+        this.leftLowerLeg = BipedKneesModel.makeHalfLeg(this, modelSizeIn, 0.0F, -6.0F, -2.0F, 0, 22, true);
         this.rightLeg.addChild(this.rightLowerLeg);
         this.leftLeg.addChild(this.leftLowerLeg);
     }