+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog].
 
 ## [v1.2.1-1.16.5] - 2021-07-25
+### Added
+- Player animations, mostly taken from [Console Experience] mod
+- Eating animation for both eating and drinking
+- Rowing animation when in a boat
+- Riding animation when on a horses
+- Animation for inspecting hand held items, such as compass and clock
 ### Changed
 - Renamed zombie knees element to humanoid knees, now also includes piglin-like mobs
+- Made humanoid knees look better when walking and riding
 
 ## [v1.2-1.16.5] - 2021-07-24
 ### Added
@@ -64,3 +71,4 @@ The format is based on [Keep a Changelog].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Puzzles Lib]: https://www.curseforge.com/minecraft/mc-mods/puzzles-lib
+[Console Experience]: https://www.curseforge.com/minecraft/mc-mods/console-hud